+++ backend/canisters/event_relay/api/src/updates/mod.rs
@@ -1,1 +1,2 @@
 pub mod push_events;
+pub mod push_events_v2;