+++ backend/canisters/event_relay/impl/src/jobs/get_transactions.rs
@@ -59,14 +59,14 @@ async fn run_async() -> Milliseconds {
             };
             events.push(
                 EventBuilder::new("mint", timestamp)
-                    .with_source(ledger_canister_id.to_string())
+                    .with_source(ledger_canister_id.to_string(), false)
                     .with_json_payload(&payload)
                     .build(),
             );
             if mint.to == treasury_account {
                 events.push(
                     EventBuilder::new("transfer_to_treasury", timestamp)
-                        .with_source(ledger_canister_id.to_string())
+                        .with_source(ledger_canister_id.to_string(), false)
                         .with_json_payload(&payload)
                         .build(),
                 );
@@ -79,14 +79,14 @@ async fn run_async() -> Milliseconds {
             };
             events.push(
                 EventBuilder::new("burn", timestamp)
-                    .with_source(ledger_canister_id.to_string())
+                    .with_source(ledger_canister_id.to_string(), false)
                     .with_json_payload(&payload)
                     .build(),
             );
             if burn.from == treasury_account {
                 events.push(
                     EventBuilder::new("transfer_from_treasury", timestamp)
-                        .with_source(ledger_canister_id.to_string())
+                        .with_source(ledger_canister_id.to_string(), false)
                         .with_json_payload(&payload)
                         .build(),
                 );
@@ -95,7 +95,7 @@ async fn run_async() -> Milliseconds {
             if transfer.from == treasury_account {
                 events.push(
                     EventBuilder::new("transfer_from_treasury", transaction.timestamp / NANOS_PER_MILLISECOND)
-                        .with_source(ledger_canister_id.to_string())
+                        .with_source(ledger_canister_id.to_string(), false)
                         .with_json_payload(&TransactionPayload {
                             index,
                             amount: transfer.amount.0.try_into().unwrap(),
@@ -105,7 +105,7 @@ async fn run_async() -> Milliseconds {
             } else if transfer.to == treasury_account {
                 events.push(
                     EventBuilder::new("transfer_to_treasury", transaction.timestamp / NANOS_PER_MILLISECOND)
-                        .with_source(ledger_canister_id.to_string())
+                        .with_source(ledger_canister_id.to_string(), false)
                         .with_json_payload(&TransactionPayload {
                             index,
                             amount: transfer.amount.0.try_into().unwrap(),