+++ backend/canisters/group/impl/src/new_joiner_rewards.rs
@@ -83,7 +83,6 @@ fn send_reward_transferred_message(user_id: UserId, transfer: nns::CompletedCryp
             replies_to: None,
             forwarded: false,
             sender_is_bot: true,
-            sender_name_override: None,
             correlation_id: 0,
             now: state.env.now(),
         },