+++ backend/canisters/local_group_index/api/src/updates/push_events_v2.rs
@@ -1,0 +1,2 @@
+pub type Args = event_store_canister::PushEventsArgs;
+pub type Response = ();