+++ backend/canisters/local_user_index/api/src/updates/push_events.rs
@@ -1,2 +1,4 @@
-pub type Args = event_store_canister::PushEventsArgs;
+use event_store_canister::PushEventsArgsPrevious;
+
+pub type Args = PushEventsArgsPrevious;
 pub type Response = ();