+++ backend/canisters/local_user_index/impl/src/updates/push_events.rs
@@ -1,12 +1,39 @@
 use crate::guards::caller_is_local_user_canister;
 use crate::{mutate_state, RuntimeState};
 use canister_tracing_macros::trace;
+use event_store_canister::{Anonymizable, IdempotentEvent};
 use ic_cdk_macros::update;
-use local_user_index_canister::push_events::*;
+use local_user_index_canister::push_events::Args as ArgsPrevious;
+use local_user_index_canister::push_events_v2::Args;
 
 #[update(guard = "caller_is_local_user_canister")]
 #[trace]
-fn push_events(args: Args) {
+fn push_events(args: ArgsPrevious) {
+    mutate_state(|state| {
+        push_events_impl(
+            Args {
+                events: args
+                    .events
+                    .into_iter()
+                    .filter(|e| e.name != "message_sent")
+                    .map(|e| IdempotentEvent {
+                        idempotency_key: e.idempotency_key,
+                        name: e.name,
+                        timestamp: e.timestamp,
+                        user: e.user.map(|u| Anonymizable::new(u, true)),
+                        source: e.source.map(|s| Anonymizable::new(s, true)),
+                        payload: e.payload,
+                    })
+                    .collect(),
+            },
+            state,
+        )
+    })
+}
+
+#[update(guard = "caller_is_local_user_canister")]
+#[trace]
+fn push_events_v2(args: Args) {
     mutate_state(|state| push_events_impl(args, state))
 }