+++ backend/canisters/user/impl/src/updates/start_video_call.rs
@@ -81,7 +81,6 @@ pub fn handle_start_video_call(
         replies_to: None,
         forwarded: false,
         sender_is_bot: true,
-        sender_name_override: None,
         correlation_id: 0,
         now,
     };