+++ backend/canisters/group/CHANGELOG.md
@@ -18,6 +18,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Check start video call permission in access_token ([#5524](https://github.com/open-chat-labs/open-chat/pull/5524))
 - Update `event_store` packages to latest version ([#5535](https://github.com/open-chat-labs/open-chat/pull/5535))
+- Anonymize all Group canisters in events ([#5568](https://github.com/open-chat-labs/open-chat/pull/5568))
 
 ## [[2.0.1096](https://github.com/open-chat-labs/open-chat/releases/tag/v2.0.1096-group)] - 2024-03-08